Add some interactive games (sort of) from Ch 10
There are two games, and one "game", from Chapter 10 of Graham
Hutton's "Programming in Haskell":
 - hangman
 - nim
 - the game of life

The game of life is implemented in a straightforward way, since
it's purely functional after the initial setting (which is hardcoded
in anyway) so the interactive part is just a display.

But nim and hangman both involve user interaction, and I used the
`mtl` `MonadState` functionality to combine a state monad with an
IO monad (to be honest it wouldn't be that hard to thread the state
by hand and just use IO, but I want to get better at monad
transformers).

This commit concludes Chapter 10, although I may also put the
tic-tac-toe from Chapter 11 into this project.