Add human vs human tic-tac-toe
This is from Chapter 11 of Graham Hutton's book.

I think I'm going to put off doing an AI player for now, since
brute-forcing the game trees doesn't seem all that exciting,
and there's a recall on the book so I want to make sure I work
on the tiny abstract machine exercises.